Add qla2xxx_wwn WWN type
The qla2xxx_wwn type is used in policy/fabric_qla2xxx.lio but is not a
valid type in rtslib.config.Config.validate_val(). This commit adds
validation for this type in rtslib.config.Config.validate_val() and
rtslib.utils.is_valid_wwn().

This allows a system with existing targets (created by an older
targetcli/lio-utils) to be validated by rtslib and for the configuration
to be saved.

Signed-off-by: Chris Boot <bootc@bootc.net>